clearer prints, change type

+++ JorPNG/zlib.c
@@ -37,8 +37,8 @@ void process_zlib_stream(uint8_t* data, uint32_t length, Bitstream* output) {
 
   zlib_stream.ADLER32 = read_bytes(sizeof(zlib_stream.ADLER32), &bitstream);
 
-  uint32_t bitcount = (bitstream.length - bitstream.byte_position) * 8 - bitstream.bit_position;
-  printf("%u/%u bits processed (%u left)\n", ((bitstream.length * 8) - bitcount), bitstream.length * 8, bitcount);
+  size_t bitcount = (bitstream.length - bitstream.byte_position) * 8 - bitstream.bit_position;
+  printf("%llu/%llu bits processed (%llu left)\n", ((bitstream.length * 8) - bitcount), bitstream.length * 8, bitcount);
 
   print_stream_info(&zlib_stream);
 }
